Eliminate dependency of printk_patcher_test on runtime package, reduce import loops
