+++ pkg/ebpf/testdata/c/logdebug-test.c
